+++ stdlib/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "stdlib"
 version = "0.1.0"
-edition = "2018"
+edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
